📍 split dtos by resolver endpoints
