+++ src/dtos/payment/get-payment/PaymentFindUniqueArgs.ts
@@ -1,5 +1,5 @@
 import { ArgsType, Field } from '@nestjs/graphql';
-import { PaymentWhereUniqueInput } from '@dtos/payment/PaymentWhereUniqueInput';
+import { PaymentWhereUniqueInput } from '@/dtos/payment/get-payment/PaymentWhereUniqueInput';
 
 @ArgsType()
 class PaymentFindUniqueArgs {
