+++ src/dtos/payment/get-payments/PaymentWhereInput.ts
@@ -1,10 +1,10 @@
 import { Field, InputType } from '@nestjs/graphql';
 import { Type } from 'class-transformer';
 import { IsOptional } from 'class-validator';
-import { DateTimeNullableFilter } from '@util/DateTimeNullableFilter';
-import { FloatNullableFilter } from '@util/FloatNullableFilter';
-import { StringFilter } from '@util/StringFilter';
-import { StringNullableFilter } from '@util/StringNullableFilter';
+import { DateTimeNullableFilter } from '@/util/DateTimeNullableFilter';
+import { FloatNullableFilter } from '@/util/FloatNullableFilter';
+import { StringFilter } from '@/util/StringFilter';
+import { StringNullableFilter } from '@/util/StringNullableFilter';
 
 @InputType()
 class PaymentWhereInput {
