+++ src/modules/payment.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { PaymentResolver } from '@resolvers/payment.resolver';
-import { HelperService } from '@services/helper.service';
-import { PaymentService } from '@services/payment.service';
-import { PrismaService } from '@services/prisma.service';
 import { PubSub } from 'graphql-subscriptions';
+import { PaymentResolver } from '@/resolvers/payment.resolver';
+import { HelperService } from '@/services/helper.service';
+import { PaymentService } from '@/services/payment.service';
+import { PrismaService } from '@/services/prisma.service';
 
 @Module({
   providers: [
