+++ src/services/helper.service.ts
@@ -1,3 +1,6 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
 export class HelperService {
   public randomNumBetween(min: number, max: number): number {
     return Math.floor(Math.random() * (max - min)) + min;
